+++ test/utest-online/testlinkapi_generic_online_test.py
@@ -386,7 +386,11 @@ class TestLinkAPIOnlineTestCase(unittest.TestCase):
                                         buildname='build 4713', 
                                         platformname='platform 4714',
                                         user='username',action='unassignOne') 
-                                  
+    
+    def test_getProjectKeywords_unknownID(self):
+        with self.assertRaisesRegex(TLResponseError, '7000.*4711'):
+            self.client.getProjectKeywords(4711) 
+                                 
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']
     unittest.main()